fix: update onboarding flow to handle redirects and improve logout process
